services/backend/requirements.txt
celery==5.2.3
channels==3.0.4
channels-redis==3.4.0
daphne==3.0.2
Django==3.2.12
djangorestframework==3.13.1
django-celery-beat==2.2.1
django-celery-results==2.3.0
djoser==2.1.0
drf-yasg==1.20.0
flake8==4.0.1
flower==1.0.0
gunicorn==20.1.0
isort==5.10.1
numpy==1.22.3
opencv-python-headless==4.5.5.64
Pillow==9.0.1
psycopg2-binary==2.9.2
python-decouple==3.6
redis==4.1.4
scipy==1.8.0
whitenoise==6.0.0
